refactoring setup, show, and tally
