+++ src/vtp/cli/tally_contests.py
@@ -25,11 +25,64 @@ See 'tally_contests.py -h' for usage information.
 """
 
 # Global imports
+import argparse
+import re
 import sys
 
 # Local imports
+from vtp.core.common import Common
 from vtp.ops.tally_contests_operation import TallyContestsOperation
 
+
+def parse_arguments(argv):
+    """Parse arguments from a command line or from the constructor"""
+
+    parser = argparse.ArgumentParser(
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        description="""
+Will tally all the contests so far merged to the main branch and
+report the results.  The results are computed on a voting center basis
+(git submodule) basis.
+
+Note - the current implementation relies on git submodules (individual
+git repos) to break up the tally data of an election.  If there is
+only one git repository and the election is large, then a potentiallu
+large amount of memory will be used in executing the tallies.  One
+short term fix for this is to limit the number of contests being
+tallied.
+
+Also note that the current implementation does not yet support
+tallying across git submodules/repos.
+""",
+    )
+
+    Common.add_election_data_dir(parser)
+    parser.add_argument(
+        "-c",
+        "--contest_uid",
+        default="",
+        help="limit the tally to a specific contest uid",
+    )
+    parser.add_argument(
+        "-t",
+        "--track_contests",
+        default="",
+        help="a comma separated list of contests checks to track",
+    )
+    Common.add_verbosity(parser)
+    parsed_args = parser.parse_args(argv)
+
+    # Validate required args
+    if parsed_args.track_contests:
+        if not bool(re.match("^[0-9a-f,]", parsed_args.track_contests)):
+            raise ValueError(
+                "The track_contests parameter only accepts a comma separated (no spaces) "
+                "list of contest checks/digests to track."
+            )
+        parsed_args.track_contests = parsed_args.track_contests.split(",")
+    else:
+        parsed_args.track_contests = []
+    return parsed_args
 
 def main():
     """
@@ -40,9 +93,19 @@ def main():
     source file.
     """
 
+    # Parse args
+    parsed_args = parse_arguments(sys.argv)
+
     # do it
-    tco = TallyContestsOperation(sys.argv[1:])
-    tco.run()
+    tco = TallyContestsOperation(
+        parsed_args.election_data_dir,
+        parsed_args.verbosity,
+        parsed_args.printonly,
+        )
+    tco.run(
+        contest_uid=parsed_args.contest_uid,
+        track_contests=parsed_args.track_contests,
+        )
 
 
 # If called directly via this file
