+++ src/vtp/ops/show_contests_operation.py
@@ -23,10 +23,8 @@ See 'show_contest.py -h' for usage information.
 """
 
 # Standard imports
-import argparse
 import logging
 import os
-import re
 
 # Local libraries
 from vtp.core.common import Common, Globals, Shellout
@@ -40,42 +38,13 @@ class ShowContestsOperation:
     description (immediately below this) in the source file.
     """
 
-    @staticmethod
-    def parse_arguments(argv):
-        """Parse arguments from a command line or from the constructor"""
-
-        safe_args = Common.cast_thing_to_list(argv)
-        parser = argparse.ArgumentParser(
-            formatter_class=argparse.RawDescriptionHelpFormatter,
-            description="""
-    will print the CVRs (Cast Vote Records) for the supplied contest(s)
-    """,
-        )
-        Common.add_election_data(parser)
-        parser.add_argument(
-            "-c",
-            "--contest-check",
-            help="a comma separate list of contests digests to validate/display",
-        )
-
-        Common.add_verbosity(parser)
-        Common.add_printonly(parser)
-        parsed_args = parser.parse_args(safe_args)
-
-        # Validate required args
-        Common.verify_election_data(parsed_args)
-        if not parsed_args.contest_check:
-            raise ValueError("The contest check is required")
-        if not bool(re.match("^[0-9a-f,]", parsed_args.contest_check)):
-            raise ValueError(
-                "The contest_check parameter only accepts a comma separated (no spaces) "
-                "list of contest checks/digests to track."
-            )
-        return parsed_args
-
-    def __init__(self, unparsed_args):
+    def __init__(self, election_data_dir: str, verbosity: int, printonly: bool):
         """Only to module-ize the scripts and keep things simple and idiomatic."""
-        self.parsed_args = ShowContestsOperation.parse_arguments(unparsed_args)
+        self.election_data_dir = election_data_dir
+        self.verbosity = verbosity
+        self.printonly = printonly
+        # Configure logging
+        Common.configure_logging(verbosity)
 
     def validate_digests(self, digests, election_data_dir, error_digests):
         """Will scan the supplied digests for validity.  Will print and
@@ -92,7 +61,7 @@ class ShowContestsOperation:
                         "--batch-check=%(objectname) %(objecttype)",
                         "--buffer",
                     ],
-                    verbosity=self.parsed_args.verbosity,
+                    verbosity=self.verbosity,
                     input=input_data,
                     text=True,
                     check=True,
@@ -123,14 +92,11 @@ class ShowContestsOperation:
     # main
     ################
     # pylint: disable=duplicate-code
-    def run(self):
+    def run(self, contest_check: str = ""):
         """Main function - see -h for more info"""
 
-        # Configure logging
-        Common.configure_logging(self.parsed_args.verbosity)
-
         # Create a VTP ElectionData object if one does not already exist
-        the_election_config = ElectionConfig.configure_election()
+        the_election_config = ElectionConfig.configure_election(self.election_data_dir)
 
         # Check the ElectionData
         election_data_dir = os.path.join(
@@ -141,11 +107,11 @@ class ShowContestsOperation:
         # First validate the digests
         error_digests = set()
         self.validate_digests(
-            self.parsed_args.contest_check, election_data_dir, error_digests
+            contest_check, election_data_dir, error_digests
         )
         valid_digests = [
             digest
-            for digest in self.parsed_args.contest_check.split(",")
+            for digest in contest_check.split(",")
             if digest not in error_digests
         ]
         # show/log the digests
@@ -155,18 +121,18 @@ class ShowContestsOperation:
 
 # For future reference just in case . . .
 # this is a loop of shell commands
-#        for digest in self.parsed_args.contest_check.split(','):
+#        for digest in contest_check.split(','):
 #            if digest not in error_digests:
 #                Shellout.run(['git', 'log', '-1', digest], check=True)
 
 # this does not work well enough either
-#        input_data = '\n'.join(self.parsed_args.contest_check.split(',')) + '\n'
+#        input_data = '\n'.join(contest_check.split(',')) + '\n'
 #        Shellout.run(
 #            ['git', 'cat-file', '--batch=%(objectname)'],
 #            input=input_data,
 #            text=True,
 #            check=True,
-#            verbosity=self.parsed_args.verbosity)
+#            verbosity=self.verbosity)
 
 # End Of Class
