+++ src/vtp/ops/tally_contests_operation.py
@@ -25,9 +25,7 @@ See 'tally_contests.py -h' for usage information.
 """
 
 # Standard imports
-import argparse
 import logging
-import re
 
 # Local imports
 from vtp.core.ballot import Ballot
@@ -45,88 +43,34 @@ class TallyContestsOperation:
     description (immediately below this) in the source file.
     """
 
-    @staticmethod
-    def parse_arguments(argv):
-        """Parse arguments from a command line or from the constructor"""
-
-        safe_args = Common.cast_thing_to_list(argv)
-        parser = argparse.ArgumentParser(
-            formatter_class=argparse.RawDescriptionHelpFormatter,
-            description="""
-Will tally all the contests so far merged to the main branch and
-report the results.  The results are computed on a voting center basis
-(git submodule) basis.
-
-Note - the current implementation relies on git submodules (individual
-git repos) to break up the tally data of an election.  If there is
-only one git repository and the election is large, then a potentiallu
-large amount of memory will be used in executing the tallies.  One
-short term fix for this is to limit the number of contests being
-tallied.
-
-Also note that the current implementation does not yet support
-tallying across git submodules/repos.
-""",
-        )
-
-        Common.add_election_data(parser)
-        parser.add_argument(
-            "-c",
-            "--contest_uid",
-            default="",
-            help="limit the tally to a specific contest uid",
-        )
-        parser.add_argument(
-            "-t",
-            "--track_contests",
-            default="",
-            help="a comma separated list of contests checks to track",
-        )
-        parser.add_argument(
-            "-x",
-            "--do_not_pull",
-            action="store_true",
-            help="Before tallying the votes, pull the ElectionData repo",
-        )
-        Common.add_verbosity(parser)
-        parsed_args = parser.parse_args(safe_args)
-
-        # Validate required args
-        Common.verify_election_data(parsed_args)
-        if parsed_args.track_contests:
-            if not bool(re.match("^[0-9a-f,]", parsed_args.track_contests)):
-                raise ValueError(
-                    "The track_contests parameter only accepts a comma separated (no spaces) "
-                    "list of contest checks/digests to track."
-                )
-            parsed_args.track_contests = parsed_args.track_contests.split(",")
-        else:
-            parsed_args.track_contests = []
-        return parsed_args
-
-    def __init__(self, unparsed_args):
+    def __init__(self, election_data_dir: str, verbosity: int, printonly: bool):
         """Only to module-ize the scripts and keep things simple and idiomatic."""
-        self.parsed_args = TallyContestsOperation.parse_arguments(unparsed_args)
+        self.election_data_dir = election_data_dir
+        self.verbosity = verbosity
+        self.printonly = printonly
+        # Configure logging
+        Common.configure_logging(verbosity)
 
-    ################
-    # main
-    ################
     # pylint: disable=duplicate-code
-    def run(self):
+    def run(
+        self,
+        contest_uid: str = "",
+        track_contests: str = "",
+        ):
         """Main function - see -h for more info"""
 
         # Configure logging
-        Common.configure_logging(self.parsed_args.verbosity)
+        Common.configure_logging(self.verbosity)
 
         # Create a VTP ElectionData object if one does not already exist
-        the_election_config = ElectionConfig.configure_election()
+        the_election_config = ElectionConfig.configure_election(self.election_data_dir)
 
         # git pull the ElectionData repo so to get the latest set of
         # remote CVRs branches
         a_ballot = Ballot()
         with Shellout.changed_cwd(a_ballot.get_cvr_parent_dir(the_election_config)):
             Shellout.run(
-                ["git", "pull"], verbosity=self.parsed_args.verbosity, check=True
+                ["git", "pull"], verbosity=self.verbosity, check=True
             )
 
         # Will process all the CVR commits on the main branch and tally
@@ -145,10 +89,10 @@ tallying across git submodules/repos.
         # everything in a separate loop.
         for contest_batch in sorted(contest_batches):
             # Maybe skip
-            if self.parsed_args.contest_uid != "":
+            if contest_uid != "":
                 if (
                     contest_batches[contest_batch][0]["CVR"]["uid"]
-                    != self.parsed_args.contest_uid
+                    != contest_uid
                 ):
                     continue
             # Create a Tally object for this specific contest
@@ -166,7 +110,7 @@ tallying across git submodules/repos.
             #        import pdb; pdb.set_trace()
             try:
                 the_tally.tallyho(
-                    contest_batches[contest_batch], self.parsed_args.track_contests
+                    contest_batches[contest_batch], track_contests
                 )
                 # Print stuff
                 the_tally.print_results()
